Fixed Cube boundary generation method
There are two boundary generation methods in VacuumMesher, Cube and Bounding Box.
Previous to this commit, when using a Cube boundary, the method being called to generate the faces
was not tesselating them, returning a cubic boundary comprised of just the edges of the cube. When the
individual boundary faces were being combined to create the full cubic boundary, this was resulting in a#
horrific mess of a boundary. I have adjusted the genBoundary method to correctly call GenTriangulatedSquare instead of
just genSquare now, which should fix #18